# api/openapi.yaml
openapi: 3.0.3
info:
  title: Omnifolio API
  version: 1.0.0
  description: Personal investments tracker.
servers:
  - url: /api
    description: Local dev (proxied via Vite or Caddy)
...
paths:
...
  /instruments:
    get:
      tags: [instruments]
      summary: List instruments with optional search and pagination
      operationId: listInstruments
      parameters:
        - name: q
          in: query
          required: false
          description: Filter by ticker or name (case-insensitive substring).
          schema:
            type: string
            minLength: 1
            maxLength: 50
        - name: assetClass
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AssetClass"
        - name: scope
          in: query
          required: false
          description: |
            Restrict results to a scope.
            `mine` — caller's personal instruments only.
            `global` — canonical catalog only.
            Omitted — union of both (caller's personal + global).
          schema:
            type: string
            enum: [mine, global]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated list of instruments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstrumentList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [instruments]
      summary: Create a personal instrument (real estate, vehicle, other property)
      operationId: createInstrument
      description: |
        Creates a personal instrument owned by the caller. Restricted to manual asset
        classes (`real_estate`, `vehicle`, `other_asset`); exchange-traded and cash
        classes are global-only and maintained by the cron worker.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonalInstrumentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Conflict:
      description: Resource conflict (e.g., duplicate)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: RFC 7807 Problem Details
      required: [title, status]
      properties:
        type:
          type: string
          example: /errors/validation
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          example: 422
        detail:
          type: string
        instance:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
...
    AssetClass:
      type: string
      enum: [ru_stock, ru_bond, ru_etf, us_stock, us_etf, crypto, cash, real_estate, vehicle, other_asset]
    InstrumentScope:
      type: string
      description: |
        `global` — canonical catalog row maintained by cron (exchange-traded / cash classes).
        `personal` — row owned by the calling user (manual asset classes: real_estate, vehicle, other_asset).
      enum: [global, personal]
    Instrument:
      type: object
      required: [id, ticker, assetClass, currency, name, scope, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        ticker:
          type: string
        assetClass:
          $ref: "#/components/schemas/AssetClass"
        currency:
          type: string
          pattern: "^[A-Z]{3,5}$"
        name:
          type: string
        scope:
          $ref: "#/components/schemas/InstrumentScope"
        userId:
          type: string
          format: uuid
          nullable: true
          description: Owner user id. Null for global rows, set for personal rows.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        currentPrice:
          type: string
          description: Last known price as decimal string. Omitted if no price recorded yet.
          example: "123.45"
        priceUpdatedAt:
          type: string
          format: date-time
          description: When currentPrice was last fetched. Omitted if no price recorded yet.
    InstrumentList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Instrument"
        total:
          type: integer
          minimum: 0
          description: Total instruments matching the filter (for pagination).
    CreatePersonalInstrumentRequest:
      type: object
      required: [ticker, name, assetClass, currency, initialPrice]
      properties:
        ticker:
          type: string
          pattern: "^[A-Za-z0-9_-]{1,32}$"
          description: Short label / identifier (user-visible).
        name:
          type: string
          minLength: 1
          maxLength: 100
        assetClass:
          type: string
          enum: [real_estate, vehicle, other_asset]
          description: Personal-only asset classes. Exchange/cash classes are rejected.
        currency:
          type: string
          pattern: "^[A-Z]{3,5}$"
        initialPrice:
          type: string
          description: Positive decimal as string, e.g. "12000000".
          example: "12000000"
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
security:
  - sessionCookie: []