# apps/api/internal/storage/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package storage

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
	"github.com/shopspring/decimal"
)

type Account struct {
	ID             uuid.UUID
	UserID         uuid.UUID
	SourceType     string
	Name           string
	LastSyncedAt   pgtype.Timestamptz
	LastSyncStatus *string
	LastSyncError  *string
	CreatedAt      pgtype.Timestamptz
	UpdatedAt      pgtype.Timestamptz
}

type AccountCredential struct {
	AccountID  uuid.UUID
	Ciphertext []byte
	Nonce      []byte
	KeyVersion int32
	CreatedAt  pgtype.Timestamptz
	UpdatedAt  pgtype.Timestamptz
}

type Deposit struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	Month     pgtype.Date
	Amount    decimal.Decimal
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type FxRate struct {
	Date    pgtype.Date
	FromCcy string
	ToCcy   string
	Rate    decimal.Decimal
}

type Instrument struct {
	ID         uuid.UUID
	Ticker     string
	AssetClass string
	Currency   string
	Name       string
	CreatedAt  pgtype.Timestamptz
	UpdatedAt  pgtype.Timestamptz
	UserID     uuid.NullUUID
}

type InstrumentExternalID struct {
	Source       string
	NativeID     string
	InstrumentID uuid.UUID
	CreatedAt    pgtype.Timestamptz
}

type PortfolioSnapshot struct {
	UserID          uuid.UUID
	SnapshotDate    pgtype.Date
	DisplayCurrency string
	GrandTotal      decimal.Decimal
	ByAssetClass    []byte
	ByCurrency      []byte
	ByAccount       []byte
	CreatedAt       pgtype.Timestamptz
}

type Position struct {
	AccountID    uuid.UUID
	InstrumentID uuid.UUID
	Quantity     decimal.Decimal
	UpdatedAt    pgtype.Timestamptz
}

type Price struct {
	InstrumentID uuid.UUID
	Price        decimal.Decimal
	FetchedAt    pgtype.Timestamptz
}

type Session struct {
	TokenHash  []byte
	UserID     uuid.UUID
	CreatedAt  pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
	LastSeenAt pgtype.Timestamptz
}

type User struct {
	ID              uuid.UUID
	Email           string
	PasswordHash    string
	DisplayCurrency string
	CreatedAt       pgtype.Timestamptz
	UpdatedAt       pgtype.Timestamptz
}